+++ app/src/main/java/com/example/lab1/network/Downloader.java
@@ -20,11 +20,11 @@ public class Downloader {
     public String setWeatherURL(String[] coordinates) {
 
         /*---------- Test DATA -------------*/
-        mWeatherUrl = "https://maceo.sth.kth.se/weather/forecast?lonLat=lon/" + coordinates[0] + "/lat/" + coordinates[1];
+        //mWeatherUrl = "https://maceo.sth.kth.se/weather/forecast?lonLat=lon/" + coordinates[0] + "/lat/" + coordinates[1];
         //mWeatherUrl = "https://maceo.sth.kth.se/weather/forecast?lonLat=lon/14.333/lat/60.383";
 
         /*---------- REAL DATA -------------*/
-        //mWeatherUrl = "https://opendata-download-metfcst.smhi.se/api/category/pmp3g/version/2/geotype/point/lon/" + coordinates[0] + "/lat/" + coordinates[1] + "/data.json";
+        mWeatherUrl = "https://opendata-download-metfcst.smhi.se/api/category/pmp3g/version/2/geotype/point/lon/" + coordinates[0] + "/lat/" + coordinates[1] + "/data.json";
 
         Log.d(LOG_TAG, "Coordinate link accessed");
         return mWeatherUrl;
@@ -35,10 +35,10 @@ public class Downloader {
 
         /*---------- TEST DATA -------------*/
         //mCityUrl = "https://maceo.sth.kth.se/weather/search?location=Sigfridstorp";
-        mCityUrl = "https://maceo.sth.kth.se/weather/search?location=" + cityName;
+        //mCityUrl = "https://maceo.sth.kth.se/weather/search?location=" + cityName;
 
         /*---------- REAL DATA -------------*/
-        //mCityUrl = "https://www.smhi.se/wpt-a/backend_solr/autocomplete/search/" + cityName;
+        mCityUrl = "https://www.smhi.se/wpt-a/backend_solr/autocomplete/search/" + cityName;
 
         Log.d(LOG_TAG, "Weather link accessed");
         return mCityUrl;
