+++ app/src/main/res/layout/activity_main.xml
@@ -16,99 +16,40 @@
         app:headerLayout="@layout/menu_header"
         app:menu="@menu/menu" />
 
-    <androidx.constraintlayout.widget.ConstraintLayout
+    <LinearLayout
         android:layout_width="match_parent"
-        android:layout_height="match_parent">
+        android:layout_height="0dp"
+        app:layout_constraintTop_toTopOf="parent"
+        android:orientation="vertical">
 
         <include
-            android:id="@+id/main_toolbar"
+            android:id="@+id/toolbar_layout"
             layout="@layout/toolbar"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-        <TextView
-            android:id="@+id/approvedtime_view"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="8dp"
-            android:text="@string/approvedTime"
-            android:textSize="11sp"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toEndOf="@+id/message_view"
-            app:layout_constraintTop_toBottomOf="@id/main_toolbar" />
-
-        <TextView
-            android:id="@+id/message_view"
-            android:layout_width="250dp"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="8dp"
-            android:text="@string/header"
-            android:textAppearance="@android:style/TextAppearance.DeviceDefault.Medium"
-            app:layout_constraintBottom_toTopOf="@+id/recycler_view"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/main_toolbar" />
-
-
-        <AutoCompleteTextView
-            android:id="@+id/editText_city_input"
-            android:layout_width="100dp"
-            android:layout_height="40dp"
-            android:layout_marginStart="16dp"
-            android:layout_marginTop="8dp"
-            android:layout_marginBottom="8dp"
-            android:ems="10"
-            android:hint="@string/input_city_hint"
-            android:inputType="textPersonName"
-            android:textAppearance="@style/TextAppearance.AppCompat.Small"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/recycler_view" />
-
-        <TextView
-            android:id="@+id/textView_loc"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="16dp"
-            android:text="@string/textView_loc"
-            android:textAlignment="viewStart"
-            android:textColor="@android:color/black"
-            android:textStyle="bold"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintStart_toStartOf="parent" />
-
-        <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/recycler_view"
             android:layout_width="match_parent"
-            android:layout_height="0dp"
-            android:layout_margin="4dp"
-            app:layout_constraintBottom_toTopOf="@+id/set"
-            app:layout_constraintTop_toBottomOf="@+id/message_view"
-            tools:layout_editor_absoluteX="4dp" />
+            android:layout_height="wrap_content"/>
 
-
-        <TextView
-            android:id="@+id/netView"
-            android:layout_width="wrap_content"
-            android:layout_height="18dp"
-            android:text="@string/network"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toStartOf="@+id/set"
-            app:layout_constraintStart_toEndOf="@+id/editText_city_input"
-            app:layout_constraintTop_toBottomOf="@+id/recycler_view" />
-
-        <Button
-            android:id="@+id/set"
-            android:layout_width="wrap_content"
+        <LinearLayout
+            android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:layout_marginTop="8dp"
-            android:layout_marginEnd="16dp"
-            android:layout_marginBottom="8dp"
-
-            android:text="@string/set"
-            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintTop_toBottomOf="@id/toolbar_layout"
             app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/recycler_view" />
+            app:layout_constraintStart_toStartOf="parent">
 
-    </androidx.constraintlayout.widget.ConstraintLayout>
+            <TextView
+                android:id="@+id/netView"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="@string/network"
+                android:layout_weight="2" />
+
+        </LinearLayout>
+
+        <FrameLayout
+            android:id="@+id/content_container"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent" />
+
+    </LinearLayout>
+
+
 </androidx.drawerlayout.widget.DrawerLayout>
