+++ app/src/main/res/values/colors.xml
@@ -13,4 +13,5 @@
     <color name="bc_gray">#E8E8E8</color>
     <color name="red">#FF0000</color>
     <color name="blue">#0000FF</color>
+    <color name='dark_red'>#3F0000</color>
 </resources>
