+++ app/src/main/res/values/strings.xml
@@ -40,6 +40,8 @@
         <item>Kiruna</item>
         <item>Haparanda</item>
     </string-array>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 
 
 </resources>
